REF: Added a service layer
