--- src/main/java/stenden/spring/HouseService.java
@@ -1,22 +1,0 @@
-package stenden.spring;
-
-import org.springframework.stereotype.Service;
-import stenden.spring.data.House;
-
-@Service
-public class HouseService {
-
-  public House getJdbcHouse(Long id) {
-    return null;
-  }
-
-  public House getHibernateHouse(Long id) {
-    return null;
-  }
-
-  public House getPureJdbcHouse(Long id) {
-    return null;
-  }
-
-}
-
