+++ src/main/java/stenden/spring/resource/HouseResource.java
@@ -6,8 +6,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import stenden.spring.data.House;
-import stenden.spring.data.HouseRepository;
-import stenden.spring.data.jpa.SpringJpaRepository;
+import stenden.spring.service.HouseService;
 
 // We're telling Spring MVC that this is a REST controller
 // This treats methods with @RequestMapping as if it had the @ResponseBody annotation
@@ -15,50 +14,38 @@ import stenden.spring.data.jpa.SpringJpaRepository;
 @RestController
 // The base request mapping this controller listens to, we respond to anything starting with /data
 @RequestMapping("/data")
-public class DataResource {
+public class HouseResource {
 
-  private final HouseRepository jdbcTemplateRepository;
-  private final HouseRepository pureJdbcRepository;
-  private final HouseRepository hibernateRepository;
-  private final HouseRepository entityManagerJpaRepository;
-  private final SpringJpaRepository springJpaRepository;
+  private final HouseService houseService;
 
   @Autowired
-  public DataResource(HouseRepository jdbcTemplateRepository,
-                      HouseRepository pureJdbcRepository,
-                      HouseRepository hibernateRepository,
-                      HouseRepository entityManagerJpaRepository,
-                      SpringJpaRepository springJpaRepository) {
-    this.jdbcTemplateRepository = jdbcTemplateRepository;
-    this.pureJdbcRepository = pureJdbcRepository;
-    this.hibernateRepository = hibernateRepository;
-    this.entityManagerJpaRepository = entityManagerJpaRepository;
-    this.springJpaRepository = springJpaRepository;
+  public HouseResource(HouseService houseService) {
+    this.houseService = houseService;
   }
 
   @GetMapping("/templatejdbc/{id}")
   public House getTemplateJdbcHouse(@PathVariable("id") Long id) {
-    return jdbcTemplateRepository.getByID(id);
+    return houseService.getTemplateJdbcHouse(id);
   }
 
   @GetMapping("/purejdbc/{id}")
   public House getPureJdbcHouse(@PathVariable("id") Long id) {
-    return pureJdbcRepository.getByID(id);
+    return houseService.getPureJdbcHouse(id);
   }
 
   @GetMapping("/hibernate/{id}")
   public House getHibernateHouse(@PathVariable("id") Long id) {
-    return hibernateRepository.getByID(id);
+    return houseService.getHibernateHouse(id);
   }
 
   @GetMapping("/jpa/{id}")
   public House getJpaHouse(@PathVariable("id") Long id) {
-    return entityManagerJpaRepository.getByID(id);
+    return houseService.getJpaHouse(id);
   }
 
   @GetMapping("/springjpa/{id}")
   public House getSpringJpaHouse(@PathVariable("id") Long id) {
-    return springJpaRepository.findById(id).get();
+    return houseService.getSpringJpaHouse(id);
   }
 
 }
